Reduce stack size to 1024 kByte to start more containers
The default stack size is 8192 kByte on a 32bit OS. This limits the maximum number of processes
we could start on a 512 MByte or 1 GByte Raspberry Pi. Therefor we can only start a limited number
of Dcoker containers.

With a reduced stack size, we are able to start an even larger number of containers.